Делит код на ES2015 модули
