+++ js/main.js
@@ -1,4 +1,7 @@
-import { photosArray } from './data.js';
-export { photosArray };
+import { generatePhotosArray } from './photos.js';
+
+const photosArray = generatePhotosArray();
 
 console.log(photosArray);
+
+export { photosArray };
