+++ js/utils.js
@@ -1,11 +1,11 @@
-// Генерация случайного числа в диапазоне
+// Генерирует случайное целое число в заданном диапазоне
 const getRandomInteger = (minValue, maxValue) => {
   const min = Math.ceil(Math.min(minValue, maxValue));
   const max = Math.floor(Math.max(minValue, maxValue));
   return Math.floor(Math.random() * (max - min + 1) + min);
 };
 
-// Получение случайного элемента массива
+// Возвращает случайный элемент из массива
 const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
 
 export { getRandomInteger, getRandomArrayElement };
